+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - Updated code to mint permalinks bound for IIIF/Cantaloupe instead of Fedora. (DR-3006)
 
+## Unreleased
+
+### Added
+- Added method to add containsUnrestrictedMaterial to solr docs. (DR-3425)
+
 ## [3.1.0] - 2025-03-12
 
 ### Added